fixes + initial version of incomeDeclarationFactory

+++ src/main/java/com/durys/jakub/socialfund/declarations/domain/IncomeGroupResolver.java
@@ -3,5 +3,5 @@ package com.durys.jakub.socialfund.declarations.domain;
 import java.math.BigDecimal;
 
 public interface IncomeGroupResolver {
-    Integer resolve(Long year, BigDecimal income);
+    Integer resolve(Integer year, BigDecimal income);
 }
